requirements.txt
fastapi==0.109.2
pydantic
uvicorn==0.27.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
pydantic-settings
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0